pyproject.toml
[project]
name = "pyzeros"
version = "0.1.0"
description = "TODO"
readme = "README.md"
license = "MIT"
authors = [
  { name = "Elian NEPPEL", email = "elian.dev@posteo.com" }
]
requires-python = ">=3.10,<3.14"
keywords = ["asyncio", "robotics", "ros2", "zenoh", "publisher", "subscriber"]

dependencies = [
"asyncio_for_robotics",
# Python-side ROS message support
"xxhash",
"cyclonedds-nightly",
"ros2_pyterfaces",
"eclipse-zenoh",
"cydr",
"numpy",
"cyclonedds_idl",
]

[build-system]
requires = ["setuptools >= 58.2.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pyzeros", "pyzeros.*"]

[tool.setuptools.package-data]
"pyzeros" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
